Extract CommunityEngagementCheckService for ex parte CE step
Move combined hours/income orchestration out of CertificationBusinessProcess
into CommunityEngagementCheckService (determine: aggregate, record on case,
publish Strata events). Rename from ExParteCommunityEngagementCheckService.

Update specs to stub the service; add dedicated service spec.

Made-with: Cursor
