+++ reporting-app/app/services/community_engagement_check_service.rb
@@ -1,0 +1,61 @@
+# frozen_string_literal: true
+
+# Called by CertificationBusinessProcess at the ex parte community engagement step.
+# Aggregates hours and income, records a combined determination on the case, and publishes
+# Strata events for transitions/notifications (see NotificationsEventListener).
+class CommunityEngagementCheckService
+  class << self
+    # @param kase [CertificationCase]
+    def determine(kase)
+      certification = Certification.find(kase.certification_id)
+      hours_data = HoursComplianceDeterminationService.aggregate_hours_for_certification(certification)
+      income_data = IncomeComplianceDeterminationService.aggregate_income_for_certification(certification)
+
+      hours_ok = hours_compliant?(hours_data)
+      income_ok = IncomeComplianceDeterminationService.compliant_for_total_income?(income_data[:total_income])
+
+      kase.record_ex_parte_ce_combined_assessment(
+        hours_data: hours_data,
+        income_data: income_data,
+        hours_ok: hours_ok,
+        income_ok: income_ok
+      )
+
+      publish_workflow_events(
+        kase: kase,
+        certification: certification,
+        hours_data: hours_data,
+        income_data: income_data,
+        hours_ok: hours_ok,
+        income_ok: income_ok
+      )
+    end
+
+    private
+
+    def hours_compliant?(hours_data)
+      hours_data[:total_hours].to_f >= HoursComplianceDeterminationService::TARGET_HOURS
+    end
+
+    def publish_workflow_events(kase:, certification:, hours_data:, income_data:, hours_ok:, income_ok:)
+      if hours_ok || income_ok
+        Strata::EventManager.publish("DeterminedHoursMet", {
+          case_id: kase.id,
+          certification_id: certification.id
+        })
+      elsif hours_data[:hours_by_source][:ex_parte].to_f.positive?
+        Strata::EventManager.publish("DeterminedHoursInsufficient", {
+          case_id: kase.id,
+          certification_id: certification.id,
+          hours_data: hours_data,
+          income_data: income_data
+        })
+      else
+        Strata::EventManager.publish("DeterminedActionRequired", {
+          case_id: kase.id,
+          certification_id: certification.id
+        })
+      end
+    end
+  end
+end
